#94 可読性を高めるためのコンポーネントの分割
